pyproject.toml
[project]
name = "gather-news"
version = "0.1.0"
description = "Tech news digest auto-generation system"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "feedparser",
    "google-genai",
    "pyyaml",
    "jinja2",
    "tzdata",
]
